core/requirements.txt
Cython==0.29.21
aiofiles==0.6.0
aioredis==1.3.1
aio-pika==6.8.0
async_lru==1.0.2
aiokafka==0.7.0
aiomysql==0.0.20
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.61.2
fastapi-utils==0.2.1
h11==0.11.0
httpx==0.17.1
idna==2.10
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pydantic==1.7.2
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.1
pytest-asyncio==0.14.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
starlette==0.13.6
tabulate==0.8.7
toml==0.10.2
urllib3==1.26.2
uvicorn==0.12.2
yoyo-migrations==7.2.0
websockets==8.1
